surface: move decl to wlr_{sub,}compositor.h
